Rearranged backend files to create better modularity
